packages/pyengineqt/pyproject.toml
[project]
name = "f8pyengineqt"
version = "0.1.0"
description = "Operator flow engine with NodeGraphQt"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "f8pysdk",
    "nodegraphqt>=0.6.43",
    "pydantic>=2.12.5",
    "pyside6>=6.10.1",
    "qtpy>=2.4.3",
    "setuptools>=80.9.0",
    "nats-py>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["f8pyengineqt"]

[tool.uv]
package = true
